添加注释

+++ src/main/java/com/hniecs/mainserver/entity/UserEntity.java
@@ -11,11 +11,13 @@ import lombok.Setter;
  * @logs[0] 2020-09-13 01:47 yijie 添加了密码生成与校验
  */
 public class UserEntity {
+    // 用户id
     @Getter
-    private int id;
-    @Getter
-    @Setter
+    private long id;
+    // 用户名
+    @Getter @Setter
     private String userName;
+    // SHA算法加密后的密码
     private String passwordSHA;
 
     public UserEntity(String userName, String password) {
@@ -34,6 +36,10 @@ public class UserEntity {
     public void setPassword (String password) {
         this.passwordSHA = Password.generatePasswordHash(password);
     }
+    /**
+     * 校验密码是否正确
+     * @param password 待校验密码
+     */
     public boolean vertifyPWD (String password) {
         return Password.checkPasswordHash(this.passwordSHA, password);
     }
